refactor(map): extract stop icon and popup helpers from useMapLayers
Extract two coherent rendering concerns out of useMapLayers.ts into
dedicated modules under src/hooks/mapLayers/:

- stopIcons.ts: isValidCoordinate, StopMarkerType, createStopIcon
- popups.ts: PopupOptions, bindRichStopPopup, bindRichLinePopup

Reduces useMapLayers.ts from ~800 to 588 lines. Map behavior unchanged.

Also fix line tooltip cursor overlap by switching direction to 'top'
with sticky tracking so the tooltip always renders above the pointer.

Refs: STORY-0009
Generated-by: GitHub Copilot (Claude Sonnet 4.6)
