+++ docs/stories/STORY-0009.md
@@ -2,7 +2,7 @@
 
 ## Status
 
-OPEN
+CLOSED
 
 ## Context
 
@@ -25,5 +25,5 @@ Do this in slices; this story is first extraction step, not full redesign.
 
 ## Completion
 
-- Date:
-- Outcome:
+- Date: 2026-03-09
+- Outcome: Extracted stop icon rendering (`mapLayers/stopIcons.ts`) and popup/tooltip rendering (`mapLayers/popups.ts`) from `useMapLayers.ts`. Hook reduced from ~800 to 588 lines. Also fixed line-tooltip cursor overlap by switching to `direction: 'top'` with sticky tracking. Map behavior unchanged; all tests pass.
