+++ src/hooks/useMapLayers.ts
@@ -3,96 +3,12 @@ import L from 'leaflet';
 import { Connection, Leg, Stop } from '../types';
 import { COLORS, getGradientColor, TRANSPORT_COLORS } from '../constants';
 import { getLegStopTimes } from '../utils/dataUtils';
-
-function isValidCoordinate(
-  lat: number | undefined | null,
-  lng: number | undefined | null
-): boolean {
-  return (
-    typeof lat === 'number' &&
-    typeof lng === 'number' &&
-    !isNaN(lat) &&
-    !isNaN(lng)
-  );
-}
-
-type StopMarkerType =
-  | 'highlighted'
-  | 'source'
-  | 'target'
-  | 'intermediate-future'
-  | 'intermediate-past'
-  | 'context'
-  | 'context-subtle';
-
-function createStopIcon(
-  type: StopMarkerType,
-  isDark: boolean,
-  colorOverride?: string
-) {
-  const sizeMap: Record<StopMarkerType, number> = {
-    highlighted: 24,
-    source: 20,
-    target: 20,
-    'intermediate-future': 14,
-    'intermediate-past': 12,
-    context: 14,
-    'context-subtle': 8,
-  };
-  const size = sizeMap[type];
-
-  let html = '';
-
-  switch (type) {
-    case 'highlighted': {
-      const hColor = colorOverride || COLORS.primary;
-      html = `
-                <div class="relative w-full h-full">
-                    <div class="absolute inset-0 rounded-full animate-ping opacity-40" style="background-color: ${hColor}"></div>
-                    <div class="relative w-full h-full rounded-full border-[3px] shadow-lg flex items-center justify-center" style="background-color: ${hColor}; border-color: ${isDark ? '#0f172a' : '#fff'}">
-                        <div class="w-2 h-2 bg-white rounded-full"></div>
-                    </div>
-                </div>`;
-      break;
-    }
-
-    case 'source':
-    case 'target': {
-      // Unified "Black Dot" style for Start (Source) and End (Target)
-      // Light Mode: White Fill, Fat Black Border
-      // Dark Mode: Black Fill, Fat White Border
-      const borderColor = isDark ? '#ffffff' : '#0f172a';
-      const fillColor = isDark ? '#0f172a' : '#ffffff';
-      html = `<div class="w-full h-full rounded-full shadow-lg" style="background-color: ${fillColor}; border: 4px solid ${borderColor}; box-sizing: border-box;"></div>`;
-      break;
-    }
-
-    case 'intermediate-future':
-      html = `<div class="w-full h-full rounded-full bg-white dark:bg-slate-900 border-[3px] border-indigo-500 dark:border-indigo-400 shadow-sm"></div>`;
-      break;
-
-    case 'intermediate-past':
-      // High visibility gray for passed stops
-      html = `<div class="w-full h-full rounded-full bg-slate-400 dark:bg-slate-600 border-[2px] border-slate-600 dark:border-slate-400"></div>`;
-      break;
-
-    case 'context':
-      html = `<div class="w-full h-full rounded-full bg-white dark:bg-slate-800 border-2 border-slate-300 dark:border-slate-600 shadow-sm"></div>`;
-      break;
-
-    case 'context-subtle':
-      html = `<div class="w-full h-full rounded-full bg-slate-300/50 dark:bg-slate-600/50"></div>`;
-      break;
-  }
-
-  const wrapperHtml = `<div class="flex items-center justify-center transition-transform duration-300 hover:scale-110" style="width: ${size}px; height: ${size}px;">${html}</div>`;
-  return L.divIcon({
-    className: 'bg-transparent',
-    html: wrapperHtml,
-    iconSize: [size, size],
-    iconAnchor: [size / 2, size / 2],
-  });
-}
+import { bindRichStopPopup, bindRichLinePopup } from './mapLayers/popups';
+import {
+  isValidCoordinate,
+  createStopIcon,
+  StopMarkerType,
+} from './mapLayers/stopIcons';
 
 interface UseMapLayersProps {
   map: L.Map | null;
@@ -670,93 +586,3 @@ const drawIsolineStops = (
     }
   });
 };
-
-// Unified Popup Logic
-interface PopupOptions {
-  time?: string;
-  isCurrent?: boolean;
-  isPast?: boolean;
-  isDark: boolean;
-  timezone: string;
-  permanent?: boolean;
-  highlightColor?: string;
-  subtitle?: string;
-}
-
-function bindRichStopPopup(layer: L.Layer, stop: Stop, options: PopupOptions) {
-  if (!layer.bindTooltip) return;
-
-  const {
-    isDark,
-    isCurrent,
-    isPast,
-    highlightColor,
-    subtitle,
-    permanent,
-    time,
-    timezone,
-  } = options;
-
-  const bg = isDark
-    ? 'bg-slate-900/95 border-slate-700'
-    : 'bg-white/95 border-slate-200';
-  const textColor = isDark ? 'text-slate-100' : 'text-slate-900';
-
-  let titleColorClass = textColor;
-  let titleStyle = '';
-
-  if (isCurrent) {
-    if (highlightColor) titleStyle = `color: ${highlightColor}`;
-    else titleColorClass = 'text-indigo-600 dark:text-indigo-400';
-  } else if (isPast) {
-    titleColorClass = 'text-slate-400';
-  }
-
-  let subtitleHtml = '';
-  if (time) {
-    const date = new Date(time);
-    const timeStr = date.toLocaleTimeString('en-GB', {
-      hour: '2-digit',
-      minute: '2-digit',
-      hour12: false,
-      timeZone: timezone,
-    });
-    subtitleHtml = `<div class="text-[10px] font-mono font-bold mt-0.5 ${isPast ? 'text-slate-400 line-through decoration-slate-400/50' : 'text-slate-600 dark:text-slate-300'}">${timeStr}</div>`;
-  } else if (subtitle) {
-    subtitleHtml = `<div class="text-[10px] font-mono font-bold mt-0.5" style="color:${highlightColor || 'inherit'}">${subtitle}</div>`;
-  }
-
-  const content = `
-        <div class="font-sans text-center min-w-[60px]">
-            <div class="font-bold text-xs ${titleColorClass}" style="${titleStyle}">${stop.name}</div>
-            ${subtitleHtml}
-            ${isCurrent && !permanent ? '<div class="text-[9px] font-bold text-slate-400 uppercase tracking-widest mt-1">Selected</div>' : ''}
-        </div>`;
-
-  layer.bindTooltip(content, {
-    permanent,
-    direction: 'top',
-    offset: permanent ? [0, -12] : [0, -10],
-    className: `!${bg} backdrop-blur-md border shadow-xl rounded-lg px-2 py-1.5`,
-  });
-}
-
-function bindRichLinePopup(
-  layer: L.Layer,
-  leg: Leg,
-  isDark: boolean,
-  isPast: boolean
-) {
-  if (!layer.bindTooltip) return;
-  const bg = isDark
-    ? 'bg-slate-900/95 border-slate-700'
-    : 'bg-white/95 border-slate-200';
-  const text = isDark ? 'text-slate-100' : 'text-slate-900';
-  const subText = isDark ? 'text-slate-400' : 'text-slate-500';
-  const content = `<div class="flex items-center gap-3 p-1 font-sans min-w-[140px]"><div class="flex flex-col"><div class="flex items-center gap-1.5 mb-0.5"><span class="font-bold text-xs ${text} px-1.5 py-0.5 rounded bg-slate-100 dark:bg-slate-800">${leg.trip?.route.shortName || 'Route'}</span><span class="text-[10px] ${subText}">to</span><span class="font-bold text-xs ${text}">${leg.trip?.headSign}</span></div>${isPast ? `<div class="text-[10px] font-bold text-slate-400 uppercase tracking-wider">Already departed</div>` : `<div class="text-[10px] font-mono ${subText}">Next Stop: ${leg.toStop?.name}</div>`}</div></div>`;
-  layer.bindTooltip(content, {
-    sticky: true,
-    direction: 'center',
-    className: `!${bg} backdrop-blur-md border shadow-xl rounded-xl px-0 py-0 overflow-hidden`,
-  });
-}
